.devcontainer/src/bot/handlers/commands.py
from telegram import Update
from telegram.ext import ContextTypes, CommandHandler

async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    await update.message.reply_text("🚀 Bot is running!")

def register_commands(app):
    app.add_handler(CommandHandler("start", start))
